Make scheduler interval compile-time configurable
Automatically calculate the reload value for the RTC wake up timer to
make it easier for the developer to re-configure the interval at which
the core will be woken up and the scheduler run. The default value of
10 milliseconds can be overriden with the BC_SCHEDULER_INTERVAL_MS
macro defined in bc_scheduler.h.

A scheduler interval can also be provided on the make command line as
follows:

make SCHEDULER_INTERVAL=100 debug

Make sure to run make clean before you change the scheduler interval
value on the command line.