Add functions to configure RTC write protection
Refactor the code to enable/disable RTC write protection so that it
can be called from user/application code. Add nested-invocation guards
so that write unlock can be safely called repeatedly.

The functions are defined inline in the header file to make sure that
only a minimum amount of overhead is added to the caller.