Clear RTC RSF flag when waking up from deep sleep
RTC shadow registers are not updated when SysClock is not running,
e.g., in a deep sleep mode. Clear the RTC RSF bit when waking up from a
deep sleep mode to let the application know that it may need to wait up
to two RTCClock cycles before reading the registers.