Faster functions to get and set RTC date and time
This patch provides a faster implementation of bc_rtc_get_date_time and
bc_rtc_set_date_time. The new functions use memoization, pre-computed
tables, and a couple of C-language tricks to speed the conversion up.
Function bc_rtc_get_datetime runs in under 100 microseconds in the
worst case and in 33 microseconds in the best case.

Both functions detect the mode of the RTC (12-hour or 24-hour format).

Both functions use the standard C language structure tm to pass broken
down date and time values back and forth. This is potentially a
breaking API change that may affect other git repositories or
applications. Nevertheless, the originall implementation was very slow,
so it's probably worth it.