Add fast function to retrieve UNIX time from RTC
The new function is called bc_rtc_get_timestamp and it runs in under
100 microseconds at 2.1 MHz. The function has been optimized for speed.
It is meant to be called from the scheduler when waking up from deep
sleep or Standby mode with SysClk turned off.

The function returns the UNIX time value represented by the standard
struct timespec value.

In order for the function to work correctly, the RTC must be in 24-hour
format, must use UTC date and time, and the year must be between 2000
and 2099.