Explicitly switch RTC into 24-hour mode
The RTC peripheral supports two modes: 24-hour and 12-hour (am/pm).
I originally thought that the functions that interact with the RTC
would auto-detect the mode to work correctly. Now, I think it's better
to explicitly configure the RTC peripheral in 24-hour mode on the first
initialization. Having the RTC in 24-hour format simplifies the
implementation of those functions and improves their run time.

It is easy to convert from the 24-hour format to the 12-hour format in
application software if needed.