Implement application_idle weak function
Modify sleep-related code (the semaphore) from the scheduler to the
sleep manager and create a weak function called application_idle that
will be invoked whenever the scheduler has no work to do. The function
symbol is declared as weak so that the function can be overriden from
the application. The default implementation only invokes bc_sleep.

The application_idle task allows the application developer to customize
the behavior of the system when there is no work to do, i.e., before
entering a low power mode, without the need to modify the SDK. This
could be used, e.g., to disable peripherals or to manually activate the
Standby mode.