Only include .d files when building
There is no need to generate and include .d when the user runs one of
the targets that do not build anything, e.g., clean. Not generating
and not including dependency files in such cases allows one run targets
such as clean even in cases when there is something seriously wrong with
one of the depencency files generated in previous steps.