Add a Python script to generate firmware version from git (#310)
* Add a Python script to generate firmware version from git

This commit adds a simple Python script that generates command line
options for gcc with macro definitions that contain a version string
and build date obtained from git version history. The output looks
as follows:

-D GIT_VERSION="v1.1.1-3-ga2c587fc" -D BUILD_DATE="2022-01-26 16:03:25 UTC"

The script is meant to be invoked from platformio.ini in the
corresponding firmware repository as follows:

[env:debug]
build_flags =
    !python .pio/libdeps/debug/twr-sdk/git-version.py

[env:release]
build_flags =
    -D RELEASE
    !python .pio/libdeps/release/twr-sdk/git-version.py

* Add preprocessor macro VERSION

The string in VERSION will be set to the most recent tag or to "vdev" if
no tag was found. Git commit hash abbreviations are not used.