Make AT$AT both an action and a setter
Previously, the only supported AT$AT syntax was AT$AT=AT... This can be
a bit confusing since, technically, the AT command is not a setter. It
just passes its argument to the LoRa modem. This patch adds support for
the following syntax:

AT$AT AT+....

That is, the AT$AT prefix and the actual AT command to be passed to the
LoRa modem are space-separated. The original setter syntax is also
retained.