Pass custom AT responses from LoRa to upstream
When an application invokes a custom LoRa command via AT$AT, pass the
response generated by the LoRa module to the upstream ATCI, on separate
lines prefixed with "$LORA: ". This feature provides applications using
the Tower Core module ATCI with full access to the LoRa module's ATCI in
a "tunneled" mode.

Also, only generate an "OK" response to AT$AT once the tunneled AT
command has been actually sent to LoRa module. If it cannot be sent,
e.g., because the command could not fit into an internal buffer, send an
"ERROR" instead.